+++ app/mocks/db.mock.ts
@@ -1,4 +1,4 @@
-import { SearchLectures } from '../store/querys/user';
+import { SearchLectures } from '../store/querys/lecture';
 import { TakenLectures } from '../business/services/lecture/taken-lecture.query';
 import { CreditResponse, ResultCategoryDetailResponse } from '../business/services/result/result.type';
 import {
