+++ app/store/querys/lecture.ts
